+++ uart.h
@@ -19,7 +19,7 @@ extern uint8_t uartMemory[ADI_UART_BIDIR_MEMORY_SIZE];
 extern uint32_t uart_hErrors;
 extern ADI_UART_HANDLE uartDevice;
 extern uint8_t* rx_buffer;
-extern uint8_t rx_buffer_size;
+extern uint32_t rx_buffer_size;
 
 void uartCallback(void* pAppHandle, uint32_t nEvent, void* pArg);
 
@@ -36,7 +36,7 @@ uint8_t uartRead();
 //reads len bytes from buffer. Returns 0 in case of success and 1 otherwise
 //buf: pointer where the bytes are going to be copied to.
 //len: number of bytes to be read
-int uartReadBuffer(uint8_t* buf, uint8_t len);
+int uartReadBuffer(uint8_t* buf, uint32_t len);
 
 //writes one byte (blocking)
 //byte: data to be written
@@ -45,7 +45,7 @@ ADI_UART_RESULT uartWrite(uint8_t byte);
 //writes len bytes (blocking)
 //buffer: pointer with data to be written
 //len: number of bytes to be written
-ADI_UART_RESULT uartWriteBuffer(uint8_t* buffer, uint8_t len);
+ADI_UART_RESULT uartWriteBuffer(uint8_t* buffer, uint32_t len);
 
 //writes one byte (non-blocking)
 //byte: data to be written
@@ -54,6 +54,6 @@ ADI_UART_RESULT async_uartWrite(uint8_t byte);
 //writes len bytes (non-blocking)
 //buffer: pointer with data to be written
 //len: number of bytes to be written
-ADI_UART_RESULT async_uartWriteBuffer(uint8_t* buffer, uint8_t len);
+ADI_UART_RESULT async_uartWriteBuffer(uint8_t* buffer, uint32_t len);
 
 #endif /* UART_H_ */
